Aplicacao de boas praticas e flash messages

+++ src/Controller/VideoRemoveController.php
@@ -2,23 +2,44 @@
 
 namespace Alura\Mvc\Controller;
 
+use Alura\Mvc\Helper\FlashMessageTrait;
 use Alura\Mvc\Repository\VideoRepository;
+use Nyholm\Psr7\Response;
+use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Server\RequestHandlerInterface;
 
-
-class VideoRemoveController implements Controller
+class VideoRemoveController implements RequestHandlerInterface
 {
+
+    use FlashMessageTrait;
+
     public function __construct(private VideoRepository $repository)
     {
     }
 
-    public function processaRequisicao(): void
+    public function handle(ServerRequestInterface $request): ResponseInterface
     {
-        $id = filter_input(INPUT_GET, 'id', FILTER_VALIDATE_INT);
+        $query = $request->getQueryParams();
+        $id = filter_var($query['id'], FILTER_VALIDATE_INT);
+
+        if($id === null || $id === false){
+            $this->addErrorMessage('Video não encontrado.');
+            return new Response(302, [
+                'Location' => '/'
+            ]);
+        }
 
         if($this->repository->removeVideo($id) === true){
-            header("Location: /?sucesso=1");
+            $this->addSucessMessage('Video Removido com sucesso.');
+            return new Response(302, [
+                'Location' => '/'
+            ]);
         }else {
-            header("Location: /?sucesso=0");
+            $this->addErrorMessage('Remove failed');
+            return new Response(302, [
+                'Location' => '/'
+            ]);
         }
     }
 }
